project refactor and website project addition
