+++ src/App.css
@@ -371,11 +371,14 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
   margin: 14px auto 30px auto;
   text-align: center;
   width: 56%;
-  padding-top: 30px;
+  padding-top: 3vh;
 }
 .projectcard p {
   font-size: 15px;
   color: #ffffff;
+}
+.project-first{
+  padding-bottom: 5vh;
 }
 .project .nav.nav-pills {
   width: 100%;
@@ -385,7 +388,7 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
   overflow: hidden;
 }
 .project .nav.nav-pills .nav-item {
-  width: 16.66666%;
+  width: 20%;
 }
 .project .nav.nav-pills .nav-link {
   background-color: transparent;
@@ -437,13 +440,23 @@ nav.navbar .navbar-toggler[aria-expanded="true"] .navbar-toggler-icon {
   border-bottom: 1px solid rgba(255, 255, 255, 0.5);
   border: 1px solid rgba(255, 255, 255, 0.5);
 }
-.nav-link#projects-tabs-tab-fifth {
+.nav-link#projects-tabs-tab-secondlast {
   border-top: 1px solid rgba(255, 255, 255, 0.5);
   border-bottom: 1px solid rgba(255, 255, 255, 0.5);
 }
-.nav-link#projects-tabs-tab-sixth {
+.nav-link#projects-tabs-tab-last {
   border: 1px solid rgba(255, 255, 255, 0.5);
   border-radius: 0 55px 55px 0;
+}
+.project .nav.nav-pills-custom {
+  width: 100%;
+  margin: 0 auto;
+  border-radius: 50px;
+  background-color: rgb(255 255 255 / 10%);
+  overflow: hidden;
+}
+.project .nav.nav-pills-custom .nav-item {
+  width: 25%;
 }
 .proj-imgbx {
   position: relative;
